Add files via upload
For each sea state j=1 through 7, the computed u(x,t) and \delta(x,t) are included. Note that this are on a time grid coarser than the one used in computation. The time and space grid is also included, as are the initial conditions. The notations are identical as in the paper "Phase Resolved Simulation of the Landau--Alber Stability~Bifurcation".